# external/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// ScooterData defines model for ScooterData.
type ScooterData struct {
	Charge int    `json:"charge"`
	Id     string `json:"id"`
	ZoneId string `json:"zone_id"`
}

// TariffZone defines model for TariffZone.
type TariffZone struct {
	DefaultDeposit int    `json:"default_deposit"`
	Id             string `json:"id"`
	PricePerMinute int    `json:"price_per_minute"`
	PriceUnlock    int    `json:"price_unlock"`
}

// UserProfile defines model for UserProfile.
type UserProfile struct {
	HasSubscribtion bool   `json:"has_subscribtion"`
	Id              string `json:"id"`
	Trusted         bool   `json:"trusted"`
}

// PostClearMoneyForOrderJSONBody defines parameters for PostClearMoneyForOrder.
type PostClearMoneyForOrderJSONBody struct {
	Amount  *int    `json:"amount,omitempty"`
	OrderId *string `json:"order_id,omitempty"`
	UserId  *string `json:"user_id,omitempty"`
}

// PostHoldMoneyForOrderJSONBody defines parameters for PostHoldMoneyForOrder.
type PostHoldMoneyForOrderJSONBody struct {
	Amount  *int    `json:"amount,omitempty"`
	OrderId *string `json:"order_id,omitempty"`
	UserId  *string `json:"user_id,omitempty"`
}

// GetScooterDataParams defines parameters for GetScooterData.
type GetScooterDataParams struct {
	// Id Идентификатор самоката
	Id string `form:"id" json:"id"`
}

// GetTariffZoneDataParams defines parameters for GetTariffZoneData.
type GetTariffZoneDataParams struct {
	// Id Идентификатор тарифной зоны
	Id string `form:"id" json:"id"`
}

// PostUnholdMoneyForOrderJSONBody defines parameters for PostUnholdMoneyForOrder.
type PostUnholdMoneyForOrderJSONBody struct {
	OrderId *string `json:"order_id,omitempty"`
	UserId  *string `json:"user_id,omitempty"`
}

// GetUserProfileParams defines parameters for GetUserProfile.
type GetUserProfileParams struct {
	// Id Идентификатор пользователя
	Id string `form:"id" json:"id"`
}

// PostClearMoneyForOrderJSONRequestBody defines body for PostClearMoneyForOrder for application/json ContentType.
type PostClearMoneyForOrderJSONRequestBody PostClearMoneyForOrderJSONBody

// PostHoldMoneyForOrderJSONRequestBody defines body for PostHoldMoneyForOrder for application/json ContentType.
type PostHoldMoneyForOrderJSONRequestBody PostHoldMoneyForOrderJSONBody

// PostUnholdMoneyForOrderJSONRequestBody defines body for PostUnholdMoneyForOrder for application/json ContentType.
type PostUnholdMoneyForOrderJSONRequestBody PostUnholdMoneyForOrderJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Списать деньги за поездку
	// (POST /clear-money-for-order)
	PostClearMoneyForOrder(w http.ResponseWriter, r *http.Request)
	// Получить конфиги системы (динамические конфиги)
	// (GET /configs)
	GetConfigs(w http.ResponseWriter, r *http.Request)
	// Заморозить депозит на карте клиента
	// (POST /hold-money-for-order)
	PostHoldMoneyForOrder(w http.ResponseWriter, r *http.Request)
	// Получить данные по самокату (заряд и id зоны)
	// (GET /scooter-data)
	GetScooterData(w http.ResponseWriter, r *http.Request, params GetScooterDataParams)
	// Получить данные по тарифной зоне (депозит, стоимость разблокировки и минуты)
	// (GET /tariff-zone-data)
	GetTariffZoneData(w http.ResponseWriter, r *http.Request, params GetTariffZoneDataParams)
	// Разморозить депозит на карте клиента
	// (POST /unhold-money-for-order)
	PostUnholdMoneyForOrder(w http.ResponseWriter, r *http.Request)
	// Получить параметры пользователя (наличие подписки, обязательность депозита)
	// (GET /user-profile)
	GetUserProfile(w http.ResponseWriter, r *http.Request, params GetUserProfileParams)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Списать деньги за поездку
// (POST /clear-money-for-order)
func (_ Unimplemented) PostClearMoneyForOrder(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получить конфиги системы (динамические конфиги)
// (GET /configs)
func (_ Unimplemented) GetConfigs(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Заморозить депозит на карте клиента
// (POST /hold-money-for-order)
func (_ Unimplemented) PostHoldMoneyForOrder(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получить данные по самокату (заряд и id зоны)
// (GET /scooter-data)
func (_ Unimplemented) GetScooterData(w http.ResponseWriter, r *http.Request, params GetScooterDataParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получить данные по тарифной зоне (депозит, стоимость разблокировки и минуты)
// (GET /tariff-zone-data)
func (_ Unimplemented) GetTariffZoneData(w http.ResponseWriter, r *http.Request, params GetTariffZoneDataParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Разморозить депозит на карте клиента
// (POST /unhold-money-for-order)
func (_ Unimplemented) PostUnholdMoneyForOrder(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получить параметры пользователя (наличие подписки, обязательность депозита)
// (GET /user-profile)
func (_ Unimplemented) GetUserProfile(w http.ResponseWriter, r *http.Request, params GetUserProfileParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// PostClearMoneyForOrder operation middleware
func (siw *ServerInterfaceWrapper) PostClearMoneyForOrder(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostClearMoneyForOrder(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetConfigs operation middleware
func (siw *ServerInterfaceWrapper) GetConfigs(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetConfigs(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostHoldMoneyForOrder operation middleware
func (siw *ServerInterfaceWrapper) PostHoldMoneyForOrder(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostHoldMoneyForOrder(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetScooterData operation middleware
func (siw *ServerInterfaceWrapper) GetScooterData(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetScooterDataParams

	// ------------- Required query parameter "id" -------------

	if paramValue := r.URL.Query().Get("id"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "id"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "id", r.URL.Query(), &params.Id)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetScooterData(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTariffZoneData operation middleware
func (siw *ServerInterfaceWrapper) GetTariffZoneData(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTariffZoneDataParams

	// ------------- Required query parameter "id" -------------

	if paramValue := r.URL.Query().Get("id"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "id"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "id", r.URL.Query(), &params.Id)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTariffZoneData(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostUnholdMoneyForOrder operation middleware
func (siw *ServerInterfaceWrapper) PostUnholdMoneyForOrder(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostUnholdMoneyForOrder(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetUserProfile operation middleware
func (siw *ServerInterfaceWrapper) GetUserProfile(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetUserProfileParams

	// ------------- Required query parameter "id" -------------

	if paramValue := r.URL.Query().Get("id"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "id"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "id", r.URL.Query(), &params.Id)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetUserProfile(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/clear-money-for-order", wrapper.PostClearMoneyForOrder)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/configs", wrapper.GetConfigs)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/hold-money-for-order", wrapper.PostHoldMoneyForOrder)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/scooter-data", wrapper.GetScooterData)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/tariff-zone-data", wrapper.GetTariffZoneData)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/unhold-money-for-order", wrapper.PostUnholdMoneyForOrder)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/user-profile", wrapper.GetUserProfile)
	})

	return r
}